# Integrating line-log and changed-path Bloom filters